# openapi/support-service.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from identity POST /api/v1/auth/login. Never commit real tokens."
      },
      "idempotencyKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key"
      },
      "requestId": {
        "type": "apiKey",
        "in": "header",
        "name": "x-request-id"
      },
      "traceId": {
        "type": "apiKey",
        "in": "header",
        "name": "x-trace-id"
      },
      "userId": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id"
      },
      "userRoles": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-roles"
      }
    }
  },
  "info": {
    "title": "NexaTech Support Service",
    "description": "Ticket hỗ trợ và hội thoại",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/admin/support/tickets": {
      "get": {
        "operationId": "AdminSupport_listTickets_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v1/admin/support/tickets/{ticketId}": {
      "get": {
        "operationId": "AdminSupport_getTicket_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v1/admin/support/tickets/{ticketId}/assign": {
      "post": {
        "operationId": "AdminSupport_assign_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v1/admin/support/tickets/{ticketId}/messages": {
      "post": {
        "operationId": "AdminSupport_addMessage_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v1/admin/support/tickets/{ticketId}/priority": {
      "patch": {
        "operationId": "AdminSupport_updatePriority_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v1/admin/support/tickets/{ticketId}/transition": {
      "post": {
        "operationId": "AdminSupport_transition_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v1/support/tickets": {
      "get": {
        "operationId": "Tickets_list_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-tickets"
        ]
      },
      "post": {
        "operationId": "Tickets_create_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "support-tickets"
        ]
      }
    },
    "/api/v1/support/tickets/{ticketId}": {
      "get": {
        "operationId": "Tickets_get_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-tickets"
        ]
      }
    },
...
    "/api/v1/support/tickets/{ticketId}/messages": {
      "post": {
        "operationId": "Tickets_addMessage_v1",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "support-tickets"
        ]
      }
    },
    "/api/v2/admin/support/tickets": {
      "get": {
        "operationId": "AdminSupport_listTickets_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v2/admin/support/tickets/{ticketId}": {
      "get": {
        "operationId": "AdminSupport_getTicket_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v2/admin/support/tickets/{ticketId}/assign": {
      "post": {
        "operationId": "AdminSupport_assign_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v2/admin/support/tickets/{ticketId}/messages": {
      "post": {
        "operationId": "AdminSupport_addMessage_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v2/admin/support/tickets/{ticketId}/priority": {
      "patch": {
        "operationId": "AdminSupport_updatePriority_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v2/admin/support/tickets/{ticketId}/transition": {
      "post": {
        "operationId": "AdminSupport_transition_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-support"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": ""
      }
    },
    "/api/v2/support/tickets": {
      "get": {
        "operationId": "Tickets_list_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-tickets"
        ]
      },
      "post": {
        "operationId": "Tickets_create_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "support-tickets"
        ]
      }
    },
    "/api/v2/support/tickets/{ticketId}": {
      "get": {
        "operationId": "Tickets_get_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-tickets"
        ]
      }
    },
...
    "/api/v2/support/tickets/{ticketId}/messages": {
      "post": {
        "operationId": "Tickets_addMessage_v2",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-roles",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-roles",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "support-tickets"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3012",
      "description": "support-service direct (local)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Kong Gateway (local)"
    },
    {
      "url": "http://{host}:{port}",
      "description": "Kubernetes / origin placeholder (override host+port)",
      "variables": {
        "host": {
          "default": "catalog-service.nexatech.svc.cluster.local"
        },
        "port": {
          "default": "3012"
        }
      }
    },
    {
      "url": "https://api.example.invalid",
      "description": "Protected public placeholder ONLY — do not use for Try it out"
    }
  ],
...
}